[#70]feat: deviceWidth 상태 추가 및 파일 명 변경
